requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.7
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
httpie==2.4.0
idna==3.2
Markdown==3.3.4
psycopg2==2.9.1
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
PySocks==1.7.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.6
